Implement SWAR find-last

+++ cbits/find-first-byte.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-ptrdiff_t find_first_byte (uint8_t* ba, int off, int len, int w8) {
+ptrdiff_t find_first_byte (uint8_t* ba, size_t off, size_t len, int w8) {
   void* s = &(ba[off]);
   uint8_t* res = memchr (s, w8, len);
   if (res == NULL) {
